move rain muffler data provider to its own class
