+++ src/main/java/com/fouristhenumber/utilitiesinexcess/common/blocks/BlockRainMuffler.java
@@ -7,9 +7,7 @@ import java.util.List;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockContainer;
 import net.minecraft.block.material.Material;
-import net.minecraft.client.Minecraft;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
@@ -21,13 +19,7 @@ import net.minecraft.world.World;
 import com.fouristhenumber.utilitiesinexcess.common.tileentities.TileEntityRainMuffler;
 import com.fouristhenumber.utilitiesinexcess.config.blocks.BlockConfig;
 
-import cpw.mods.fml.common.Optional;
-import mcp.mobius.waila.api.IWailaConfigHandler;
-import mcp.mobius.waila.api.IWailaDataAccessor;
-import mcp.mobius.waila.api.IWailaDataProvider;
-
-@Optional.Interface(iface = "mcp.mobius.waila.api.IWailaDataProvider", modid = "Waila")
-public class BlockRainMuffler extends BlockContainer implements IWailaDataProvider {
+public class BlockRainMuffler extends BlockContainer {
 
     public BlockRainMuffler() {
         super(Material.sponge);
@@ -90,40 +82,5 @@ public class BlockRainMuffler extends BlockContainer implements IWailaDataProvid
                 StatCollector
                     .translateToLocalFormatted("tile.rain_muffler.desc.1", BlockConfig.rainMuffler.rainMufflerRange));
         }
-    }
-
-    @Override
-    public List<String> getWailaBody(ItemStack itemStack, List<String> currentTip, IWailaDataAccessor accessor,
-        IWailaConfigHandler config) {
-        EntityPlayer player = Minecraft.getMinecraft().thePlayer;
-        NBTTagCompound playerNBT = player.getEntityData();
-        NBTTagCompound persistentNBT = playerNBT.getCompoundTag(EntityPlayer.PERSISTED_NBT_TAG);
-        if (persistentNBT.getBoolean(NBT_RAIN_MUFFLED)) {
-            currentTip.add(StatCollector.translateToLocal("tile.rain_muffler.waila.global_enabled"));
-        } else {
-            currentTip.add(StatCollector.translateToLocal("tile.rain_muffler.waila.global_disabled"));
-        }
-
-        return currentTip;
-    }
-
-    // Snubs
-    public NBTTagCompound getNBTData(EntityPlayerMP player, TileEntity te, NBTTagCompound tag, World world, int x,
-        int y, int z) {
-        return tag;
-    }
-
-    public ItemStack getWailaStack(IWailaDataAccessor accessor, IWailaConfigHandler config) {
-        return null;
-    }
-
-    public List<String> getWailaHead(ItemStack itemStack, List<String> currentTip, IWailaDataAccessor accessor,
-        IWailaConfigHandler config) {
-        return currentTip;
-    }
-
-    public List<String> getWailaTail(ItemStack itemStack, List<String> currentTip, IWailaDataAccessor accessor,
-        IWailaConfigHandler config) {
-        return currentTip;
     }
 }
